gdk4_x11

Function x11_get_xatom_name_for_display

pub fn x11_get_xatom_name_for_display(
    display: &X11Display,
    xatom: Atom,
) -> Option<GString>

Returns the name of an X atom for its display. This function is meant mainly for debugging, so for convenience, unlike XAtomName() and the result doesn’t need to be freed.

Deprecated since 4.18

display

the gdk::Display where @xatom is defined

xatom

an X atom

Returns

name of the X atom; this string is owned by GDK, so it shouldn’t be modified or freed.