Struct gdk4_x11::X11DeviceXI2Builder

pub struct X11DeviceXI2Builder { /* fields omitted */ }

A builder for generating a X11DeviceXI2.

Implementations

impl X11DeviceXI2Builder

pub fn new() -> Self

Create a new X11DeviceXI2Builder.

pub fn build(self) -> X11DeviceXI2

Build the X11DeviceXI2.

pub fn device_id(self, device_id: i32) -> Self

pub fn display<P: IsA<Display>>(self, display: &P) -> Self

The gdk::Display the gdk::Device pertains to.

pub fn has_cursor(self, has_cursor: bool) -> Self

Whether the device is represented by a cursor on the screen.

pub fn name(self, name: &str) -> Self

The device name.

pub fn num_touches(self, num_touches: u32) -> Self

The maximal number of concurrent touches on a touch device.

Will be 0 if the device is not a touch device or if the number of touches is unknown.

pub fn product_id(self, product_id: &str) -> Self

Product ID of this device.

See Gdk::Device::get_product_id()``.

pub fn seat(self, seat: &Seat) -> Self

gdk::Seat of this device.

pub fn source(self, source: InputSource) -> Self

Source type for the device.

pub fn vendor_id(self, vendor_id: &str) -> Self

Vendor ID of this device.

See Gdk::Device::get_vendor_id()``.