Struct gdk4_x11::X11Display

#[repr(transparent)]
pub struct X11Display { /* private fields */ }

Implements

gdk::prelude::DisplayExt

Implementations

impl X11Display

pub fn error_trap_pop(&self) -> i32

Pops the error trap pushed by gdk_x11_display_error_trap_push(). Will XSync() if necessary and will always block until the error is known to have occurred or not occurred, so the error code can be returned.

If you don’t need to use the return value, gdk_x11_display_error_trap_pop_ignored() would be more efficient.

Returns

X error code or 0 on success

pub fn error_trap_pop_ignored(&self)

Pops the error trap pushed by gdk_x11_display_error_trap_push(). Does not block to see if an error occurred; merely records the range of requests to ignore errors for, and ignores those errors if they arrive asynchronously.

pub fn error_trap_push(&self)

Begins a range of X requests on @self for which X error events will be ignored. Unignored errors (when no trap is pushed) will abort the application. Use gdk_x11_display_error_trap_pop() or gdk_x11_display_error_trap_pop_ignored()to lift a trap pushed with this function.

pub fn default_group(&self) -> Surface

Returns the default group leader surface for all toplevel surfaces on @self. This surface is implicitly created by GDK. See gdk_x11_surface_set_group().

Returns

The default group leader surface for @self

pub fn egl_version(&self) -> Option<(i32, i32)>

Available on crate feature v4_4 only.

Retrieves the version of the EGL implementation.

Returns

true if EGL is available

major

return location for the EGL major version

minor

return location for the EGL minor version

pub fn glx_version(&self) -> Option<(i32, i32)>

Retrieves the version of the GLX implementation.

Returns

true if GLX is available

major

return location for the GLX major version

minor

return location for the GLX minor version

pub fn primary_monitor(&self) -> Monitor

Gets the primary monitor for the display.

The primary monitor is considered the monitor where the “main desktop” lives. While normal application surfaces typically allow the window manager to place the surfaces, specialized desktop applications such as panels should place themselves on the primary monitor.

If no monitor is the designated primary monitor, any monitor (usually the first) may be returned.

Returns

the primary monitor, or any monitor if no primary monitor is configured by the user

pub fn screen(&self) -> X11Screen

Retrieves the X11Screen of the @self.

Returns

the X11Screen

pub fn startup_notification_id(&self) -> GString

Gets the startup notification ID for a display.

Returns

the startup notification ID for @self

pub fn user_time(&self) -> u32

Returns the timestamp of the last user interaction on @self. The timestamp is taken from events caused by user interaction such as key presses or pointer movements. See gdk_x11_surface_set_user_time().

Returns

the timestamp of the last user interaction

pub fn grab(&self)

Call XGrabServer() on @self. To ungrab the display again, use gdk_x11_display_ungrab().

gdk_x11_display_grab()/gdk_x11_display_ungrab() calls can be nested.

pub fn set_cursor_theme(&self, theme: Option<&str>, size: i32)

Sets the cursor theme from which the images for cursor should be taken.

If the windowing system supports it, existing cursors created with gdk::Cursor::from_name() are updated to reflect the theme change. Custom cursors constructed with gdk::Cursor::from_texture() will have to be handled by the application (GTK applications can learn about cursor theme changes by listening for change notification for the corresponding GtkSetting).

theme

the name of the cursor theme to use, or None to unset a previously set value

size

the cursor size to use, or 0 to keep the previous size

pub fn set_startup_notification_id(&self, startup_id: &str)

Sets the startup notification ID for a display.

This is usually taken from the value of the DESKTOP_STARTUP_ID environment variable, but in some cases (such as the application not being launched using exec()) it can come from other sources.

If the ID contains the string “_TIME” then the portion following that string is taken to be the X11 timestamp of the event that triggered the application to be launched and the GDK current event time is set accordingly.

The startup ID is also what is used to signal that the startup is complete (for example, when opening a window or when calling gdk_display_notify_startup_complete()).

startup_id

the startup notification ID (must be valid utf8)

pub fn set_surface_scale(&self, scale: i32)

Forces a specific window scale for all windows on this display, instead of using the default or user configured scale. This is can be used to disable scaling support by setting @scale to 1, or to programmatically set the window scale.

Once the scale is set by this call it will not change in response to later user configuration changes.

scale

The new scale value

pub fn string_to_compound_text(&self, str: &str) -> (i32, GString, i32, Vec<u8>)

Convert a string from the encoding of the current locale into a form suitable for storing in a window property.

str

a nul-terminated string

Returns

0 upon success, non-zero upon failure

encoding

location to store the encoding (to be used as the type for the property)

format

location to store the format of the property

ctext

location to store newly allocated data for the property

pub fn ungrab(&self)

Ungrab @self after it has been grabbed with gdk_x11_display_grab().

pub fn utf8_to_compound_text(
&self,
str: &str
) -> Option<(GString, i32, Vec<u8>)>

Converts from UTF-8 to compound text.

str

a UTF-8 string

Returns

true if the conversion succeeded, otherwise false

encoding

location to store resulting encoding

format

location to store format of the result

ctext

location to store the data of the result

pub fn open(display_name: Option<&str>) -> Option<Display>

Tries to open a new display to the X server given by @display_name. If opening the display fails, None is returned.

display_name

name of the X display. See the XOpenDisplay() for details.

Returns

The new display

impl X11Display

pub fn egl_display(&self) -> Option<Display>

Available on crate features v4_4 and egl only.

Retrieves the EGL display connection object for the given GDK display.

This function returns NULL if GDK is using GLX.

Returns

the EGL display object

pub fn xcursor(&self, cursor: &Cursor) -> XCursor

Returns the X cursor belonging to a gdk::Cursor, potentially creating the cursor.

Be aware that the returned cursor may not be unique to @cursor. It may for example be shared with its fallback cursor. On old X servers that don’t support the XCursor extension, all cursors may even fall back to a few default cursors.

cursor

a gdk::Cursor

Returns

an Xlib Cursor.

pub fn xrootwindow(&self) -> XWindow

Returns the root X window used by gdk::Display.

Returns

an X Window

pub unsafe fn xdisplay(&self) -> *mut Display

Available on crate feature xlib only.

Returns the X display of a gdk::Display.

Returns

an X display

pub unsafe fn xscreen(&self) -> *mut Screen

Available on crate feature xlib only.

Returns the X Screen used by gdk::Display.

Returns

an X Screen

pub unsafe fn connect_xevent<F: Fn(&Self, *mut XEvent) -> Continue + 'static>(
&self,
f: F
) -> SignalHandlerId

Available on crate feature xlib only.

The ::xevent signal is a low level signal that is emitted whenever an XEvent has been received.

When handlers to this signal return true, no other handlers will be invoked. In particular, the default handler for this function is GDK’s own event handling mechanism, so by returning true for an event that GDK expects to translate, you may break GDK and/or GTK+ in interesting ways. You have been warned.

If you want this signal handler to queue a gdk::Event, you can use gdk_display_put_event().

If you are interested in X GenericEvents, bear in mind that XGetEventData() has been already called on the event, and XFreeEventData() will be called afterwards.

xevent

a pointer to the XEvent to process

Returns

true to stop other handlers from being invoked for the event. false to propagate the event further.

pub fn set_program_class(&self, program_class: &str)

Sets the program class.

The X11 backend uses the program class to set the class name part of the WM_CLASS property on toplevel windows; see the ICCCM.

display

a gdk::Display

program_class

a string