Struct gdk4_x11::X11Surface

pub struct X11Surface(_);

Implementations

impl X11Surface

pub fn get_desktop(&self) -> u32

Gets the number of the workspace self is on.

Returns

the current workspace of self

pub fn get_group(&self) -> Option<Surface>

Returns the group this surface belongs to.

Returns

The group of this surface;

pub fn get_xid(&self) -> Window

Returns the X resource (surface) belonging to a gdk::Surface.

Returns

the ID of drawable’s X resource.

pub fn move_to_current_desktop(&self)

Moves the surface to the correct workspace when running under a window manager that supports multiple workspaces, as described in the Extended Window Manager Hints specification. Will not do anything if the surface is already on all workspaces.

pub fn move_to_desktop(&self, desktop: u32)

Moves the surface to the given workspace when running unde a window manager that supports multiple workspaces, as described in the Extended Window Manager Hints specification.

desktop

the number of the workspace to move the surface to

pub fn set_frame_sync_enabled(&self, frame_sync_enabled: bool)

This function can be used to disable frame synchronization for a surface. Normally frame synchronziation will be enabled or disabled based on whether the system has a compositor that supports frame synchronization, but if the surface is not directly managed by the window manager, then frame synchronziation may need to be disabled. This is the case for a surface embedded via the XEMBED protocol.

frame_sync_enabled

whether frame-synchronization should be enabled

pub fn set_group<P: IsA<Surface>>(&self, leader: &P)

Sets the group leader of self to be leader. See the ICCCM for details.

leader

a gdk::Surface

pub fn set_skip_pager_hint(&self, skips_pager: bool)

Sets a hint on self that pagers should not display it. See the EWMH for details.

skips_pager

true to skip pagers

pub fn set_skip_taskbar_hint(&self, skips_taskbar: bool)

Sets a hint on self that taskbars should not display it. See the EWMH for details.

skips_taskbar

true to skip taskbars

pub fn set_theme_variant(&self, variant: &str)

GTK applications can request a dark theme variant. In order to make other applications - namely window managers using GTK for themeing - aware of this choice, GTK uses this function to export the requested theme variant as _GTK_THEME_VARIANT property on toplevel surfaces.

Note that this property is automatically updated by GTK, so this function should only be used by applications which do not use GTK to create toplevel surfaces.

variant

the theme variant to export

pub fn set_urgency_hint(&self, urgent: bool)

Sets a hint on self that it needs user attention. See the ICCCM for details.

urgent

true to indicate urgenct attention needed

pub fn set_user_time(&self, timestamp: u32)

The application can use this call to update the _NET_WM_USER_TIME property on a toplevel surface. This property stores an Xserver time which represents the time of the last user input event received for this surface. This property may be used by the window manager to alter the focus, stacking, and/or placement behavior of surfaces when they are mapped depending on whether the new surface was created by a user action or is a "pop-up" surface activated by a timer or some other event.

Note that this property is automatically updated by GDK, so this function should only be used by applications which handle input events bypassing GDK.

timestamp

An XServer timestamp to which the property should be set

pub fn set_utf8_property(&self, name: &str, value: Option<&str>)

This function modifies or removes an arbitrary X11 window property of type UTF8_STRING. If the given self is not a toplevel surface, it is ignored.

name

Property name, will be interned as an X atom

value

Property value, or None to delete

pub fn lookup_for_display(
display: &X11Display,
window: Window
) -> Option<X11Surface>

Looks up the gdk::Surface that wraps the given native window handle.

display

the gdk::Display corresponding to the window handle

window

an Xlib Window

Returns

the gdk::Surface wrapper for the native window, or None if there is none.