Crate gdk4_x11

Rust GDK 4 X11 bindings

The project website is here.

Rust bindings of GDK 4’s X11 backend, part of gtk4-rs.

GDK is an intermediate layer that isolates GTK from the details of the windowing system. GDK X11 contains functions specific to the X11 backend.

Minimum supported Rust version

Currently, the minimum supported Rust version is 1.54.0.

Documentation

The Rust API Stable/Development

The C API

GTK Installation instructions

Using

We recommend using crates from crates.io, as demonstrated here.

If you want to track the bleeding edge, use the git dependency instead:

[dependencies]
gdk-wayland = { git = "https://github.com/gtk-rs/gtk4-rs.git", package = "gdk4-wayland" }

Avoid mixing versioned and git crates like this:

[dependencies]
gdk-x11 = {version = "0.1", package = "gdk4-x11"}
gdk-wayland = { git = "https://github.com/gtk-rs/gtk4-rs.git", package = "gdk4-wayland" }

Features

Feature	Description

v4_4	Enable the new APIs part of GTK 4.4

See Also

glib

gio

gsk4

gdk4

gtk4

License

The Rust bindings of gdk4-x11 are available under the MIT License, please refer to it.

Re-exports

pub use ffi;

pub use gdk;

pub use gio;

pub use glib;

pub use x11;

Structs

X11AppLaunchContext

Implements

X11DeviceManagerXI2

X11DeviceManagerXI2Builder

A builder-pattern type to construct X11DeviceManagerXI2 objects.

X11DeviceXI2

Implements

X11DeviceXI2Builder

A builder-pattern type to construct X11DeviceXI2 objects.

X11Display

Implements

X11Drag

Implements

X11GLContext

This is an Abstract Base Class, you cannot instantiate it.

X11Monitor

Implements

X11Screen

X11Surface

Implements

Enums

X11DeviceType