Function gdk4_x11::x11_get_xatom_by_name_for_display

pub fn x11_get_xatom_by_name_for_display(
    display: &X11Display,
    atom_name: impl IntoGStr
) -> Atom

Returns the X atom for a gdk::Display corresponding to @atom_name. This function caches the result, so if called repeatedly it is much faster than XInternAtom(), which is a round trip to the server each time.

display

a gdk::Display

atom_name

a string

Returns

a X atom for a gdk::Display